requirements.txt
Pillow
python-dotenv
requests
ytmusicapi